Change num_samples from a u8 to a u16
Night time scenes with lights weren't rendering properly with the max u8
value (255). I had to use num_samples in the thousands range to get good
results.

+++ src/ray_tracer.rs
@@ -23,7 +23,7 @@ pub struct RayTracer {
 
     /// The number of samples taken per pixel. Higher values mean lesser noise.
     #[clap(short = 's', long = "samples", default_value_t = 48)]
-    pub num_samples: u8,
+    pub num_samples: u16,
 
     /// The maximum number of reflections per light ray. Most scenes don't need
     /// more than 20 reflections.
